Fix "comparison between signed and unsigned integer expressions"

+++ apps/openmw/mwmp/GUIDialogList.cpp
@@ -28,7 +28,7 @@ GUIDialogList::GUIDialogList(const std::string &message, const std::vector<std::
     mButton->eventMouseButtonClick += MyGUI::newDelegate(this, &GUIDialogList::mousePressed);
 
     mMessage->setCaptionWithReplacing(message);
-    for(int i = 0; i < list.size(); i++)
+    for(size_t i = 0; i < list.size(); i++)
         mListBox->addItem(list[i]);
 
 }
